test(workflow): regression test for incremental schema-namespace migration
Add a two-boot regression test for issue #928 covering both issue and
project workflow stores: write a legacy row, run migration, write a
second legacy row, run migration again, and verify both rows land in
the namespaced target schema even though it is already non-empty.

The bug described in #928 (early return when the target schema is
non-empty) was already fixed by #1014, which dropped the row_count gate
and made each row go through insert_if_absent. This test locks in the
incremental + idempotent behavior so a future regression cannot
silently re-strand late-arriving legacy rows during rolling upgrades.

Extract repair_corrupt_project_ids and migrate_*_workflows_if_needed
together with their tests into a new registry_migration submodule so
registry.rs stays under the U-16 size budget. No call-site behavior
changes; the helpers are now pub(super) and re-imported by registry.

Closes #928
