+++ crates/harness-server/src/http/builders/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod engines;
 pub(crate) mod intake;
 pub(crate) mod registry;
+pub(crate) mod registry_migration;
 pub(crate) mod services;
 pub(crate) mod storage;
